backend/requirements.txt
absl-py==2.1.0
albucore==0.0.21
albumentations==1.4.22
altair==4.2.2
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==24.3.0
bleach==6.2.0
blinker==1.9.0
blis==1.2.0
branca==0.8.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.16.0
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
cmdstanpy==1.2.5
colorama==0.4.6
confection==0.1.5
contourpy==1.3.1
cvzone==1.6.1
cycler==0.12.1
cymem==2.0.11
Cython==3.0.11
distro==1.9.0
dlib==19.24.6
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
entrypoints==0.4
eval_type_backport==0.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==2.0.3
Flask-Cors==3.0.10
flatbuffers==24.3.25
folium==0.19.4
fonttools==4.55.3
gast==0.4.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.2
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
holidays==0.64
httpcore==1.0.7
httpx==0.28.1
huepy==1.2.1
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.4.26
Jinja2==3.1.5
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.6.17
keras==2.12.0
kiwisolver==1.4.8
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
livereload==2.7.1
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
mediapipe==0.10.11
ml-dtypes==0.4.0
murmurhash==1.0.12
narwhals==1.21.1
networkx==3.4.2
numpy==1.24.4
oauthlib==3.2.2
openai==1.69.0
opencv-contrib-python==4.9.0.80
opencv-python==4.5.5.64
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optional-django==0.3.0
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
plotly==5.24.1
pmdarima==2.0.4
preshed==3.0.9
prophet==1.1.6
protobuf==3.20.3
pyarrow==18.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.19.1
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==0.19.2
python-pptx==1.0.2
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
react==4.3.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
scikit-image==0.25.0
scikit-learn==1.5.0
scipy==1.15.1
seaborn==0.13.2
shapely==2.0.6
shellingham==1.5.4
simsimd==6.2.1
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.4.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
stanio==0.5.1
statsmodels==0.14.4
streamlit==1.41.1
streamlit-aggrid==1.0.5
streamlit_folium==0.24.0
stringzilla==3.11.1
tenacity==9.0.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.1
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.12.12
toml==0.10.2
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
typer==0.15.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vite==1.5.2
wasabi==1.1.3
watchdog==6.0.0
weasel==0.4.1
webencodings==0.5.1
Werkzeug==2.0.3
wrapt==1.14.1
XlsxWriter==3.2.0
xyzservices==2024.9.0
